# backend/src/main/java/org/conferatus/timetable/backend/dto/StudyGroupResponseDTO.java
package org.conferatus.timetable.backend.dto;


import com.fasterxml.jackson.annotation.JsonProperty;
import org.conferatus.timetable.backend.model.entity.StudyGroup;

public record StudyGroupResponseDTO(
        @JsonProperty("id") Long id,
        @JsonProperty("name") String name


) {
    public StudyGroupResponseDTO(StudyGroup group) {
        this(group.getId(),
                group.getName());
    }

}