# backend/src/main/resources/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: "timetable API"
  description: "timetable API"
  version: "1.0.0"
servers:
  - url: "https://t"
paths:
...
  /api/v1/admin/semesterplans:
    get:
      summary: "GET api/v1/admin/semesterplans"
      operationId: "getSemesterPlan"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SemesterPlanDTO"
      tags:
        - semesterplan
    post:
      summary: "POST api/v1/admin/semesterplans"
      operationId: "addSemesterPlan"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SemesterPlanDTO"
      tags:
        - semesterplan
    delete:
      summary: "DELETE api/v1/admin/semesterplans"
      operationId: "deleteSemesterPlan"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SemesterPlanDTO"
      tags:
        - semesterplan
...
  /api/v1/admin/semesterplans/subject:
    post:
      summary: "POST api/v1/admin/semesterplans/subject"
      operationId: "addSubjectPlan"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "times"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "subject_name"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "subject_type"
          in: "query"
          required: true
          schema:
            type: "string"
            enum:
              - "LECTURE"
              - "TERMINAL"
              - "PRACTICAL"
              - "LABORATORY"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SemesterPlanDTO"
      tags:
        - semesterplan
    delete:
      summary: "DELETE api/v1/admin/semesterplans/subject"
      operationId: "deleteSubjectPlan"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "subjectId"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SemesterPlanDTO"
      tags:
        - semesterplan
    get:
      summary: "GET api/v1/admin/semesterplans/subject"
      operationId: "getSubjectPlan"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "subjectId"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubjectPlanDTO"
      tags:
        - semesterplan
  /api/v1/admin/semesterplans/subject/teacher:
    post:
      summary: "POST api/v1/admin/semesterplans/subject/teacher"
      operationId: "addSubjectTeacher"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "subject_id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "teacher_id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SemesterPlanDTO"
      tags:
        - semesterplan
    delete:
      summary: "DELETE api/v1/admin/semesterplans/subject/teacher"
      operationId: "deleteSubjectTeacher"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "subject_id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SemesterPlanDTO"
      tags:
        - semesterplan
  /api/v1/admin/semesterplans/groups:
    post:
      summary: "POST api/v1/admin/semesterplans/groups"
      operationId: "linkGroup"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "group_id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SemesterPlanDTO"
      tags:
        - semesterplan
  /api/v1/admin/semesterplans/subject/all:
    get:
      summary: "GET api/v1/admin/semesterplans/subject/all"
      operationId: "getAllSubjectPlans"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/SubjectPlanDTO"
      tags:
        - semesterplan
...
  /api/v1/admin/teacher/by-id:
    get:
      summary: "GET api/v1/admin/teacher/by-id"
      operationId: "getTeacherById"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeacherResponseDTO"
      tags:
        - teacher
  /api/v1/admin/teacher/by-name:
    get:
      summary: "GET api/v1/admin/teacher/by-name"
      operationId: "getTeacherByName"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeacherResponseDTO"
      tags:
        - teacher
...
  /api/v1/admin/teacher:
    post:
      summary: "POST api/v1/admin/teacher"
      operationId: "addTeacher"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeacherResponseDTO"
      tags:
        - teacher
    put:
      summary: "PUT api/v1/admin/teacher"
      operationId: "updateTeacher"
      parameters:
        - name: "current"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "new"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
      tags:
        - teacher
    delete:
      summary: "DELETE api/v1/admin/teacher"
      operationId: "deleteTeacher"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
      tags:
        - teacher
  /api/v1/admin/teacher/wishes:
    post:
      summary: "POST api/v1/admin/teacher/wishes"
      operationId: "addTeacherWish"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherWishDto"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeacherResponseDTO"
      tags:
        - teacher
  /api/v1/admin/universities:
    post:
      summary: "POST api/v1/admin/universities"
      operationId: "createUniversity"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UniversityDto"
      tags:
        - university
  /api/v1/admin/universities/link:
    post:
      summary: "POST api/v1/admin/universities/link"
      operationId: "linkUserToUniversity"
      parameters:
        - name: "universityId"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "userId"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
      tags:
        - university
components:
  schemas:
    AudienceDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        audienceGroupCapacity:
          type: "integer"
          format: "int64"
    LessonDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        teacher:
          $ref: "#/components/schemas/SimpleTeacher"
        auditory:
          $ref: "#/components/schemas/AudienceDTO"
        groups:
          type: "array"
          items:
            $ref: "#/components/schemas/StudyGroupResponseDTO"
        day_index:
          type: "string"
          enum:
            - "MONDAY"
            - "TUESDAY"
            - "WEDNESDAY"
            - "THURSDAY"
            - "FRIDAY"
            - "SATURDAY"
            - "SUNDAY"
        time_index:
          type: "integer"
          format: "int64"
...
    TeacherWishDto:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        dayOfWeek:
          type: "string"
          enum:
            - "MONDAY"
            - "TUESDAY"
            - "WEDNESDAY"
            - "THURSDAY"
            - "FRIDAY"
            - "SATURDAY"
            - "SUNDAY"
        lessonNumber:
          type: "integer"
          format: "int64"
        priority:
          type: "integer"
          format: "int64"
    SimpleTeacher:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        teacherWishes:
          type: "array"
          items:
            $ref: "#/components/schemas/TeacherWishDto"
    SubjectPlanDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        times:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        subjectType:
          type: "string"
          enum:
            - "LECTURE"
            - "TERMINAL"
            - "PRACTICAL"
            - "LABORATORY"
        teacher:
          $ref: "#/components/schemas/SimpleTeacher"
    StudyGroupResponseDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
    SemesterPlanDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        subjectPlans:
          type: "array"
          items:
            $ref: "#/components/schemas/SubjectPlanDTO"
        studyGroups:
          type: "array"
          items:
            $ref: "#/components/schemas/StudyGroupResponseDTO"
    TeacherResponseDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        lessons:
          type: "array"
          items:
            $ref: "#/components/schemas/LessonDTO"
        teacherWishes:
          type: "array"
          items:
            $ref: "#/components/schemas/TeacherWishDto"
    UniversityDto:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
    UserDto:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        username:
          type: "string"
        universityDto:
          $ref: "#/components/schemas/UniversityDto"
...